# internal/controllers/user_controller.go
package controllers

import (
	"sample-service/internal/repository"
	"github.com/labstack/echo/v4"
	"strconv"
	"sample-service/internal/response"	
	"fmt"
	"sample-service/internal/model"
)

type UserController struct {
	repo repository.UserRepository
}

// NewUserController creates a new UserController
func NewUserController(repo repository.UserRepository) *UserController {
	return &UserController{
		repo: repo,
	}
}

// @Summary Get all users
// @Description Retrieve all users from the database
// @Accept json
// @Produce json
// @Success 200 {object} response.SuccessResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /users [get]
func (uc *UserController) GetAllUsers(ctx echo.Context) error {
	users, err := uc.repo.GetAllUsers()
	if err != nil {
		return response.JSONErrorResponse(ctx, "Failed to retrieve users", err.Error())
	}
	return response.JSONSuccessResponse(ctx, "Users retrieved successfully", users)
}

// @Summary Get user by ID
// @Description Retrieve a user by their ID
// @Accept json
// @Produce json
// @Param id path int true "User ID"
// @Success 200 {object} response.SuccessResponse
// @Failure 404 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /users/{id} [get]
func (uc *UserController) GetUserByID(ctx echo.Context) error {
	id := ctx.Param("id")
	userID, err := strconv.Atoi(id)
	if err != nil {
		return response.JSONErrorResponse(ctx, "Failed to retrieve user", "Invalid user ID")
	}

	user, err := uc.repo.GetUserByID(userID)
	if err != nil {
		return response.JSONErrorResponse(ctx, "User not found", err.Error())
	}
	return response.JSONSuccessResponse(ctx, "User retrieved successfully", user)
}

// @Summary Create a new user
// @Description Create a new user in the database
// @Accept json
// @Produce json
// @Param user body model.User true "User details"
// @Success 200 {object} response.SuccessResponse
// @Failure 400 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /users [post]
func (uc *UserController) CreateUser(ctx echo.Context) error {
	var user model.User
	if err := ctx.Bind(&user); err != nil {
		return response.JSONErrorResponse(ctx, "Invalid request body", err.Error())
	}

	newUser, err := uc.repo.CreateUser(user)
	if err != nil {
        if err.Error() == fmt.Sprintf("username '%s' already exists", user.UserName) {
            return response.JSONErrorResponse(ctx, "Username already exists", err.Error())
        }
        return response.JSONErrorResponse(ctx, "Failed to create user", err.Error())
    }

	return response.JSONSuccessResponse(ctx, "User created successfully", newUser)
}

// @Summary Update a user
// @Description Update a user in the database
// @Accept json
// @Produce json
// @Param user body model.User true "User details"
// @Success 200 {object} response.SuccessResponse	
// @Failure 400 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /users/{id} [put]
func (uc *UserController) UpdateUser(ctx echo.Context) error {
	var user model.User
	if err := ctx.Bind(&user); err != nil {
		return response.JSONErrorResponse(ctx, "Invalid request body", err.Error())
	}

	updatedUser, err := uc.repo.UpdateUser(user)
	if err != nil {
		return response.JSONErrorResponse(ctx, "Failed to update user", err.Error())
	}

	return response.JSONSuccessResponse(ctx, "User updated successfully", updatedUser)
}

// @Summary Delete a user
// @Description Delete a user from the database
// @Accept json
// @Produce json
// @Param id path int true "User ID"
// @Success 200 {object} response.SuccessResponse	
// @Failure 400 {object} response.ErrorResponse
// @Failure 500 {object} response.ErrorResponse
// @Router /users/{id} [delete]
func (uc *UserController) DeleteUser(ctx echo.Context) error {
	id, err := strconv.Atoi(ctx.Param("id"))
	if err != nil {
		return response.JSONErrorResponse(ctx, "Invalid user ID", err.Error())
	}

	deleted, err := uc.repo.DeleteUser(id)
	if err != nil {
		return response.JSONErrorResponse(ctx, "Failed to delete user", err.Error())
	}

	if !deleted {
		return response.JSONErrorResponse(ctx, "User not found", fmt.Sprintf("No user found with ID %d", id))
	}

	return response.JSONSuccessResponse(ctx, "User deleted successfully", nil)
}